requirements.txt
# Core ML and Deep Learning
tensorflow>=2.12.0
keras>=2.12.0
numpy>=1.21.0
opencv-python>=4.5.0

# GUI Framework
tkinter  # Usually comes with Python

# Web Framework
flask>=2.0.0

# Data Processing and Visualization
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0

# Image Processing
Pillow>=8.3.0

# Utilities
joblib>=1.1.0
tqdm>=4.62.0

# Optional: For better performance
# tensorflow-gpu>=2.12.0  # Uncomment if you have CUDA-compatible GPU

